PepClock first pass and thoughts
First pass at what would be a FeatureSet that implements PepClock. A
time machine that lets you send a message to your future self or future
friend with encouragement or advice for their post-now shenanigan’s.

This brings up some interesting questions on who/what posts to a
channel if Other uses end-to-end encryption.