Add a `layout` field to SET_CHAT_COMPLETE_RESULTS.
This tells the embedder whether to display the group of results as a column,
row or tile.

This patch doesn't expose that setting to the high-level feature API under
the theory that it'll be better for the framework to choose some reasonable
default in most cases. We can easily plumb that through if we like.

Also, fix `npm run test:watch` so that it doesn't spam the console w/ coverage
information.

Ref #59