Force Promise polyfill to be used always.
This is to work around a suspected crash in the native Promise implementation in
iOS 10's JSContext. Depending on whether this actually fixes the issue, we might
be able to find a narrower workaround.

It requires the iOS embedder to provide a `setTimeout` implementation along the
lines of http://stackoverflow.com/a/30843832.

Fixes #79