Teach core.other.js to install features.
Version 3.10.0 requires user agents to call SET_ACTIVE_CHANNEL with the id of
the currently active channel.

Actual feature install/uninstall is taken care of by the FeatureHost utils, so
the embedder doesn't need to do anything there. This implementation saves
installed feature URLs in experimental data on the identity. We can switch that
to the correct place when courier supports it.

Ref https://github.com/other-xyz/other-chat-web/issues/104